# src/Vibetech.Educat.Web/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Vibetech Educat API",
    "description": "API для платформы репетиторов и студентов",
    "contact": {
      "name": "Vibetech",
      "email": "support@vibetech.ru"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Profile/student/{studentId}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Получение профиля студента",
        "description": "Возвращает данные профиля студента по ID",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Профиль студента",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Студент не найден"
          }
        }
      }
    },
...
    "/api/Review/student/{studentId}": {
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "Получение отзывов, оставленных студентом",
        "description": "Возвращает список отзывов, оставленных указанным студентом",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список отзывов",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Student/profile/{studentId}": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Получение профиля студента",
        "description": "Возвращает данные профиля студента по ID",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Профиль студента",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Студент не найден"
          }
        }
      },
      "put": {
        "tags": [
          "Student"
        ],
        "summary": "Обновление профиля студента",
        "description": "Обновляет данные профиля студента",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Профиль студента успешно обновлен",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Студент не найден"
          }
        }
      }
    },
    "/api/Student/{studentId}/teachers": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Получение репетиторов студента",
        "description": "Возвращает список идентификаторов пользователей (UserId) репетиторов для указанного студента",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список идентификаторов пользователей репетиторов",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Student/{studentId}/lessons": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Получение уроков студента",
        "description": "Возвращает список уроков студента с возможностью фильтрации по датам",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Начальная дата фильтра (например, 2023-05-25T00:00:00)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-05-28T16:21:59Z"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Конечная дата фильтра (например, 2023-05-30T23:59:59)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-05-28T16:21:59Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список уроков студента",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Student/{studentId}/lessons/{lessonId}": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Получение урока студента",
        "description": "Возвращает данные одного урока студента по ID",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Данные урока",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDto"
                }
              }
            }
          },
          "404": {
            "description": "Урок не найден"
          }
        }
      }
    },
    "/api/Student/{studentId}/statistics": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Получение статистики студента",
        "description": "Возвращает статистическую информацию об уроках и репетиторах студента",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Статистика студента",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentStatisticsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentStatisticsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentStatisticsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Student/search-tutors": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Поиск репетиторов",
        "description": "Поиск репетиторов с различными фильтрами",
        "parameters": [
          {
            "name": "subjectId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "minExperience",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "minRating",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список репетиторов",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeacherProfileDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeacherProfileDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeacherProfileDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Student/send-request/{teacherId}": {
      "post": {
        "tags": [
          "Student"
        ],
        "summary": "Отправка заявки репетитору",
        "description": "Отправляет заявку на обучение выбранному репетитору",
        "parameters": [
          {
            "name": "studentId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "teacherId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Заявка успешно отправлена"
          },
          "400": {
            "description": "Ошибка при отправке заявки"
          }
        }
      }
    },
    "/api/Student/lesson/{lessonId}/attachments": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Получение вложений урока",
        "description": "Возвращает список файлов, прикрепленных к уроку",
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список вложений",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "lessonId": {
            "type": "integer",
            "format": "int32"
          },
          "teacherId": {
            "type": "integer",
            "format": "int32"
          },
          "teacherName": {
            "type": "string",
            "nullable": true
          },
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "studentName": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Дата создания отзыва",
            "format": "date",
            "example": "2025-05-28"
          }
        },
        "additionalProperties": false
      },
      "StudentProfileDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}